refactor: modularize dictool.rs into command modules
Split dictool.rs (1126 lines) into focused modules under commands/:
- dict_ops.rs: fetch, compile, compile_conn, merge, diff, info, lookup, prefix
- convert_ops.rs: convert, conn_cost
- user_dict_ops.rs: user-dict add/remove/list
- config_ops.rs: romaji-export/validate, settings-export/validate
- neural_ops.rs: neural-score, generate, speculative-decode (cfg-gated)

dictool.rs is now a thin dispatch layer (~230 lines) with CLI definitions
and command routing.
