+++ engine/crates/lex-cli/src/commands/mod.rs
@@ -1,0 +1,6 @@
+pub mod config_ops;
+pub mod convert_ops;
+pub mod dict_ops;
+#[cfg(feature = "neural")]
+pub mod neural_ops;
+pub mod user_dict_ops;
