+++ engine/crates/lex-cli/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod commands;
 pub mod dict_source;
